refactor: move the Card HTTP surface into its own module
The four card handlers lived in app/api/decks.py, scattering the Card
concept and contradicting the project's own convention (one module per
resource, each with its own Router). To read the Card HTTP surface you
had to open a file named decks.py.

Move all four card handlers — including the deck-nested
/decks/{deck_id}/cards/ routes — into app/api/cards.py with its own
ROUTER. build_app now registers route_handlers=[decks.ROUTER,
cards.ROUTER]; both mount at /api and resolve to distinct paths. The
URL tree is unchanged. Update CLAUDE.md to describe one Router per
resource.
